scenarist_service/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
names==0.3.0
protobuf==4.25.2
grpcio>=1.60.1
pika==1.3.2
websockets==12.0
pyyaml==6.0.0
